[PLT-1524]Update VAT rates (#71)
The new values:
Slovakia 20 -> 23
Finland 24 -> 25.5

+++ src/Bookkeeping/Invoice/InvoiceItem/ValueAddedTax.php
@@ -5,14 +5,14 @@ namespace Landingi\BookkeepingBundle\Bookkeeping\Invoice\InvoiceItem;
 
 final class ValueAddedTax
 {
-    private int $rate;
+    private float $rate;
 
-    public function __construct(int $rate)
+    public function __construct(float $rate)
     {
         $this->rate = $rate;
     }
 
-    public function getRate(): int
+    public function getRate(): float
     {
         return $this->rate;
     }
